LLM/requirements.txt
# LangGraph (state orchestration)
langgraph>=0.1.70

# Utility packages
python-dotenv
pyyaml
pytz
httpx>=0.27.0
openai>=1.40.0
ddgs>=1.0.0

# Web Server packages
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=12.0
